build: resolved Qt 6.10 compatibility issues.

+++ mo2_utils.cmake
@@ -197,10 +197,13 @@ function(mo2_deploy_qt)
 		--verbose 0
 		--webenginewidgets
 		--websockets
-		--networkauth
 		--openglwidgets
 		--libdir dlls
 		--no-compiler-runtime)
+
+	if(MO2_QT_VERSION VERSION_LESS "6.10.0")
+        list(APPEND args --networkauth)
+    endif()
 
 	if(${DEPLOY_NOPLUGINS})
 		list(APPEND args --no-plugins)
@@ -258,18 +261,38 @@ function(mo2_deploy_qt)
 	install(CODE "${removals}")
 
 	# === End Qt6 Fix ===
-
+	
 	if(NOT ${DEPLOY_NOPLUGINS})
 		set(qtwebengine_process_exe $<IF:$<CONFIG:Debug>,QtWebEngineProcessd.exe,QtWebEngineProcess.exe>)
 		install(CODE "
-			file(RENAME \"${bin}/dlls/${qtwebengine_process_exe}\" \"${bin}/${qtwebengine_process_exe}\")
-			file(RENAME \"${bin}/qtplugins/platforms\" \"${bin}/platforms\")
-			file(RENAME \"${bin}/qtplugins/styles\" \"${bin}/styles\")
-			file(RENAME \"${bin}/qtplugins/imageformats\" \"${bin}/dlls/imageformats\")
-			file(RENAME \"${bin}/qtplugins/tls\" \"${bin}/dlls/tls\")
-			file(REMOVE_RECURSE \"${bin}/qtplugins\")
-		")
-	endif()
+			if(EXISTS \"${bin}/dlls/${qtwebengine_process_exe}\")
+				message(STATUS \"[MO2] Moving ${qtwebengine_process_exe} to root...\")
+				file(RENAME \"${bin}/dlls/${qtwebengine_process_exe}\" \"${bin}/${qtwebengine_process_exe}\")
+			endif()
+
+			foreach(_dir platforms styles)
+				if(EXISTS \"${bin}/qtplugins/\${_dir}\")
+					if(EXISTS \"${bin}/\${_dir}\")
+						file(REMOVE_RECURSE \"${bin}/\${_dir}\")
+					endif()
+					file(RENAME \"${bin}/qtplugins/\${_dir}\" \"${bin}/\${_dir}\")
+				endif()
+			endforeach()
+
+			foreach(_dir imageformats tls)
+				if(EXISTS \"${bin}/qtplugins/\${_dir}\")
+					if(EXISTS \"${bin}/dlls/\${_dir}\")
+						file(REMOVE_RECURSE \"${bin}/dlls/\${_dir}\")
+					endif()
+					file(RENAME \"${bin}/qtplugins/\${_dir}\" \"${bin}/dlls/\${_dir}\")
+				endif()
+			endforeach()
+
+			if(EXISTS \"${bin}/qtplugins\")
+				file(REMOVE_RECURSE \"${bin}/qtplugins\")
+			endif()
+			")
+    endif()
 endfunction()
 
 #! mo2_add_lupdate : generate .ts files from the given sources
